requirements.txt
gensim==3.8.0
ipywidgets==7.5.1
jupyter-contrib-nbextensions==0.5.1
matplotlib-venn==0.11.5
nltk
pip-autoremove==0.9.1
pipdeptree==0.13.2
plotly==4.0.0
pyLDAvis==2.1.2
pylint==2.3.1
sklearn
spacy==2.1.8
scikit-surprise==1.1.1
tabulate==0.8.9